+++ tests/unit/components/table/ErrorsTable.spec.js
@@ -2,6 +2,7 @@ import { installQuasarPlugin } from '@quasar/quasar-app-extension-testing-unit-j
 import { shallowMount } from '@vue/test-utils';
 import i18nConfiguration from 'src/i18n';
 import ErrorsTable from 'src/components/table/ErrorsTable.vue';
+import PluginEvent from 'src/composables/events/PluginEvent';
 
 installQuasarPlugin();
 
@@ -9,6 +10,12 @@ jest.mock('vue-i18n', () => ({
   useI18n: () => ({
     t: (t) => t,
   }),
+}));
+
+jest.mock('src/composables/events/PluginEvent', () => ({
+  RequestEvent: {
+    next: jest.fn(),
+  },
 }));
 
 describe('Test component: ErrorsTable', () => {
@@ -44,11 +51,14 @@ describe('Test component: ErrorsTable', () => {
         editorType: 'diagram',
       });
 
-      expect(wrapper.vm.columns.length).toEqual(4);
+      expect(wrapper.vm.columns.length).toEqual(7);
       expect(wrapper.vm.columns[0].name).toEqual('severity');
       expect(wrapper.vm.columns[1].name).toEqual('component');
       expect(wrapper.vm.columns[2].name).toEqual('attribute');
-      expect(wrapper.vm.columns[3].name).toEqual('message');
+      expect(wrapper.vm.columns[3].name).toEqual('file');
+      expect(wrapper.vm.columns[4].name).toEqual('line');
+      expect(wrapper.vm.columns[5].name).toEqual('column');
+      expect(wrapper.vm.columns[6].name).toEqual('message');
     });
 
     it('should have valid columns for text page', async () => {
@@ -61,6 +71,33 @@ describe('Test component: ErrorsTable', () => {
       expect(wrapper.vm.columns[1].name).toEqual('line');
       expect(wrapper.vm.columns[2].name).toEqual('column');
       expect(wrapper.vm.columns[3].name).toEqual('message');
+    });
+  });
+
+  describe('Test function: selectComponent', () => {
+    it('should emit events', () => {
+      PluginEvent.RequestEvent.next.mockClear();
+
+      wrapper.vm.selectComponent('id_1');
+
+      expect(PluginEvent.RequestEvent.next).toHaveBeenCalledTimes(2);
+      expect(PluginEvent.RequestEvent.next.mock.calls).toEqual([
+        [{ type: 'select', ids: ['id_1'] }],
+        [{ type: 'edit', id: 'id_1' }],
+      ]);
+    });
+  });
+
+  describe('Test function: selectFile', () => {
+    it('should emit event', () => {
+      PluginEvent.RequestEvent.next.mockClear();
+
+      wrapper.vm.selectFile('path');
+
+      expect(PluginEvent.RequestEvent.next).toHaveBeenCalledTimes(1);
+      expect(PluginEvent.RequestEvent.next.mock.calls).toEqual([
+        [{ type: 'openFile', path: 'path' }],
+      ]);
     });
   });
 });